+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 4.5.0:
+* +: More settings in the panel
+* +: slimScroll in the settings, in the panel
+
 # 4.4.0:
 * +: Simplified stream item events from panel